[TODO-18] 재시도 요청 관련 함수 protocol 화 (#9)
## :dart: 목적
프로토콜 및 제약을 추가하여 다른 viewmodel에서도 공통적으로 사용 가능하게 변경 적용

## :pencil:  변경 내용
 > 변경사항을 리스트 형식으로 작성해주세요

 - ViewModelProtocol 에 새로운 제약 설정
 - RetryProtocol 추가로 재사용 요청 사용시 viewmodel 에 바로 적용

## :white_check_mark: 테스트 목록
 > 테스트 목록 작성 후 완료시 체크 해주세요
 - [ ]
 - [ ]

## :information_source: 기타
